requirements.txt
selenium>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
pylint>=2.15.0